feat: move server code to internal
