+++ internal/langserver/errors.go
