+++ internal/langserver/spa.go
